Mounted check before repaint setState  (#327)
* Mounted check before repaint setState

Adding a mounted check for the Chart widget repaint method, as it's possible for the repaint method to be calling setState while the widget has been disposed.

* chore(deps): update flutter_lints to version 6.0.0

Updates the linting package to the latest version to ensure
compatibility with current Flutter standards and benefit from
improved linting rules and fixes.